update Reading rating type to Double; add findByStatus method in ReadingRepository

+++ src/main/resources/db/migration/V2__create_tb_readings.sql
@@ -4,6 +4,6 @@ CREATE TABLE tb_readings (
     start_date DATE,
     end_date DATE,
     status ENUM('READ', 'READING', 'WANT_TO_READ', 'ABANDONED') NOT NULL,
-    rating INT CHECK (rating >= 0 AND rating <= 5),
+    rating DOUBLE CHECK (rating >= 0 AND rating <= 5),
     CONSTRAINT fk_reading_book FOREIGN KEY (book_id) REFERENCES tb_books(id) ON DELETE CASCADE
 );
